+++ torchopt/optim/meta/base.py
@@ -62,41 +62,40 @@ class MetaOptimizer:
         Args:
             loss: (torch.Tensor)
                 The loss that is used to compute the gradients to the network parameters.
-        """  # pylint: disable=line-too-long
+        """
         # Step parameter only
         for i, (param_container, new_state) in enumerate(
             zip(self.param_containers_groups, self.state_groups)
         ):
-            flattened_params_or_none: List[Optional[torch.Tensor]]
-            flattened_params_or_none, container_treedef = pytree.tree_flatten(param_container)  # type: ignore[arg-type]
-            flattened_params = cast(
-                Tuple[torch.Tensor, ...],
-                tuple(filter(torch.is_tensor, flattened_params_or_none)),
+            flat_params_or_none: List[Optional[torch.Tensor]]
+            flat_params_or_none, container_treedef = pytree.tree_flatten(param_container)  # type: ignore[arg-type]
+            flat_params: Tuple[torch.Tensor, ...] = tuple(
+                filter(torch.is_tensor, flat_params_or_none)  # type: ignore[arg-type]
             )
             grads = torch.autograd.grad(
                 loss,
-                flattened_params,
+                flat_params,
                 create_graph=True,
                 allow_unused=True,
             )
             updates, new_state = self.impl.update(
-                grads,
+                grads,  # type: ignore[arg-type]
                 new_state,
-                params=flattened_params,
+                params=flat_params,  # type: ignore[arg-type]
                 inplace=False,
             )
             self.state_groups[i] = new_state
-            flattened_new_params = apply_updates(flattened_params, updates, inplace=False)
-            new_params_iter = iter(flattened_new_params)
-            flattened_new_params_or_none = [
+            flat_new_params = apply_updates(flat_params, updates, inplace=False)  # type: ignore[arg-type]
+            new_params_iter = iter(flat_new_params)  # type: ignore[call-overload]
+            flat_new_params_or_none = [
                 next(new_params_iter)
                 if isinstance(old_param_or_none, torch.Tensor)
                 else old_param_or_none
-                for old_param_or_none in flattened_params_or_none
+                for old_param_or_none in flat_params_or_none
             ]
             new_params = cast(
                 Tuple[Dict[str, Optional[torch.Tensor]], ...],
-                pytree.tree_unflatten(container_treedef, flattened_new_params_or_none),
+                pytree.tree_unflatten(container_treedef, flat_new_params_or_none),
             )
             for container, new_param in zip(param_container, new_params):
                 container.update(new_param)
@@ -107,14 +106,13 @@ class MetaOptimizer:
         from torchopt.utils import _extract_container
 
         params_container, _ = _extract_container(net, with_buffers=False)
-        flattened_params = tuple(
+        flat_params = tuple(
             filter(
                 torch.is_tensor,  # type: ignore[arg-type]
-                # pylint: disable-next=line-too-long
-                cast(List[Optional[torch.Tensor]], pytree.tree_leaves(params_container)),  # type: ignore[arg-type]
+                pytree.tree_leaves(params_container),  # type: ignore[arg-type]
             )
         )
-        optimizer_state = self.impl.init(flattened_params)
+        optimizer_state = self.impl.init(flat_params)  # type: ignore[arg-type]
         self.param_containers_groups.append(params_container)
         self.state_groups.append(optimizer_state)